chore: Implement local state for workspace contecxt
