+++ packages/threat-composer/src/configs/events.ts
@@ -13,19 +13,5 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  ******************************************************************************************************************** */
-export * from './assumptions';
-export * from './mitigations';
-export * from './threats';
-export * from './threatFieldTypes';
-export * from './workspaces';
-export * from './entities';
-export * from './composerMode';
-export * from './application';
-export * from './architecture';
-export * from './dataflow';
-export * from './dataExchange';
-export * from './events';
-export * from './components';
-export * from './referencePacks';
-export * from './windowAPI';
-export * from './appMode';
+export const EVENT_WORKSPACE_DATA_CHANGED = 'workspace_data_changed';
+export const EVENT_WORKSPACE_CHANGED = 'workspace_changed';
