+++ packages/threat-composer/src/customTypes/appMode.ts
@@ -13,19 +13,9 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  ******************************************************************************************************************** */
-export * from './assumptions';
-export * from './mitigations';
-export * from './threats';
-export * from './threatFieldTypes';
-export * from './workspaces';
-export * from './entities';
-export * from './composerMode';
-export * from './application';
-export * from './architecture';
-export * from './dataflow';
-export * from './dataExchange';
-export * from './events';
-export * from './components';
-export * from './referencePacks';
-export * from './windowAPI';
-export * from './appMode';
+import { z } from 'zod';
+import { APP_MODE_BROWSER_EXTENSION, APP_MODE_IDE_EXTENSION } from '../configs/appMode';
+
+export const AppModeSchema = z.union([z.literal(APP_MODE_BROWSER_EXTENSION), z.literal(APP_MODE_IDE_EXTENSION)]).optional();
+
+export type AppMode = z.infer<typeof AppModeSchema>;
